TIP-268: move JobInstanceRepository
