+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/Filter/DateTimeFilter.php
@@ -4,8 +4,8 @@ namespace Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Filter;
 
 use Akeneo\Component\Batch\Job\BatchStatus;
 use Akeneo\Component\Batch\Job\JobRepositoryInterface;
+use Akeneo\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 use Pim\Bundle\CatalogBundle\ProductQueryUtility;
-use Pim\Bundle\ImportExportBundle\Entity\Repository\JobInstanceRepository;
 use Pim\Component\Catalog\Exception\InvalidArgumentException;
 use Pim\Component\Catalog\Query\Filter\FieldFilterInterface;
 use Pim\Component\Catalog\Query\Filter\Operators;
@@ -24,17 +24,17 @@ class DateTimeFilter extends AbstractFieldFilter implements FieldFilterInterface
     /** @var JobRepositoryInterface */
     protected $jobRepository;
 
-    /** @var JobInstanceRepository */
+    /** @var IdentifiableObjectRepositoryInterface */
     protected $jobInstanceRepository;
 
     /**
-     * @param JobInstanceRepository  $jobInstanceRepository
-     * @param JobRepositoryInterface $jobRepository
-     * @param array $supportedFields
-     * @param array $supportedOperators
+     * @param IdentifiableObjectRepositoryInterface $jobInstanceRepository
+     * @param JobRepositoryInterface                $jobRepository
+     * @param array                                 $supportedFields
+     * @param array                                 $supportedOperators
      */
     public function __construct(
-        JobInstanceRepository $jobInstanceRepository,
+        IdentifiableObjectRepositoryInterface $jobInstanceRepository,
         JobRepositoryInterface $jobRepository,
         array $supportedFields = [],
         array $supportedOperators = []
@@ -63,7 +63,7 @@ class DateTimeFilter extends AbstractFieldFilter implements FieldFilterInterface
                 throw InvalidArgumentException::stringExpected($field, 'filter', 'updated', gettype($value));
             }
 
-            $jobInstance = $this->jobInstanceRepository->findOneBy(['code' => $value]);
+            $jobInstance = $this->jobInstanceRepository->findOneByIdentifier($value);
             $lastCompletedJobExecution = $this->jobRepository->getLastJobExecution($jobInstance, BatchStatus::COMPLETED);
             if (null === $lastCompletedJobExecution) {
                 return $this;
