+++ src/Pim/Bundle/CatalogBundle/Resources/config/query_builders.yml
@@ -144,7 +144,7 @@ services:
     pim_catalog.doctrine.query.filter.datetime:
         class: '%pim_catalog.doctrine.query.filter.datetime.class%'
         arguments:
-            - '@pim_import_export.repository.job_instance'
+            - '@akeneo_batch.job.job_instance_repository'
             - '@akeneo_batch.job_repository'
             - ['created', 'updated']
             -
