+++ src/Pim/Bundle/DataGridBundle/Resources/config/controllers.yml
@@ -29,7 +29,7 @@ services:
             - '@request'
             - '@pim_datagrid.extension.mass_action.dispatcher'
             - '@pim_datagrid.adapter.oro_to_pim_grid_filter'
-            - '@pim_import_export.repository.job_instance'
+            - '@akeneo_batch.job.job_instance_repository'
             - '@security.token_storage'
             - '@akeneo_batch.launcher.simple_job_launcher'
             - '@oro_datagrid.datagrid.manager'
