--- src/Pim/Bundle/ImportExportBundle/Resources/config/repositories.yml
@@ -1,11 +1,0 @@
-parameters:
-    pim_import_export.repository.job_instance.class:  Pim\Bundle\ImportExportBundle\Entity\Repository\JobInstanceRepository
-
-services:
-    pim_import_export.repository.job_instance:
-        class: '%pim_import_export.repository.job_instance.class%'
-        factory_service: doctrine.orm.entity_manager
-        factory_method: getRepository
-        arguments: ['%akeneo_batch.entity.job_instance.class%']
-        tags:
-            - { name: 'pim_repository' }
