+++ src/Pim/Bundle/InstallerBundle/Resources/config/fixture_loader.yml
@@ -38,7 +38,7 @@ services:
     pim_installer.processor.denormalization.job_instance:
         class: '%pim_connector.processor.denormalization.job_instance.class%'
         arguments:
-            - '@pim_import_export.repository.job_instance' # TODO, alias should be changed and class should be moved
+            - '@akeneo_batch.job.job_instance_repository'
             - '@pim_connector.factory.job_instance'
             - '@pim_installer.updater.job_instance'
             - '@validator'
